requirements.txt
Flask
flask-mysqldb
mysql-connector-python
